Moved agglimbs into their own module.
No registration required anymore due to metaclassing in datreant.core.
