+++ CHANGELOG
@@ -1,0 +1,26 @@
+========================
+ datreant.data CHANGELOG
+========================
+
+The rules for this file:
+  * entries are sorted newest-first.
+  * summarize sets of changes - don't reproduce every git log comment here.
+  * don't ever delete anything.
+  * keep the format consistent (79 char width, M/D/Y date format) and do not
+    use tabs but use spaces for formatting
+  * accompany each entry with github issue/PR number (Issue #xyz)
+  * release numbers follow "Semantic Versioning" http://semver.org
+
+------------------------------------------------------------------------------
+??/??/16 dotsdl
+
+    * 0.6.0
+
+API Changes
+
+Enhancement
+
+Changes
+
+Fixes
+
